+++ src/helpers/githubMetadata.ts
@@ -1,7 +1,7 @@
+import { fetch } from "./fetch.js";
 import { isBoolean, isNumber, isObject, isString, isUrlString } from "./guards.js";
 import { URL } from "node:url";
 import { useLogger } from "../logger.js";
-import fetch from "cross-fetch";
 
 type RequestInit = Exclude<Parameters<typeof fetch>[1], undefined>;